+++ macros/src/codegen/idle.rs
@@ -68,7 +68,7 @@ pub fn codegen(
             #[allow(non_snake_case)]
             fn #name(#context: #name::Context) -> ! {
                 use rtic::Mutex as _;
-                use rtic::mutex_prelude::*;
+                use rtic::mutex::prelude::*;
 
                 #(#stmts)*
             }