+++ macros/src/codegen/software_tasks.rs
@@ -131,7 +131,7 @@ pub fn codegen(
                 #[allow(non_snake_case)]
                 fn #name(#context: #name::Context #(,#inputs)*) {
                     use rtic::Mutex as _;
-                    use rtic::mutex_prelude::*;
+                    use rtic::mutex::prelude::*;
 
                     #(#stmts)*
                 }