+++ src/lib.rs
@@ -42,6 +42,11 @@ pub use cortex_m_rtic_macros::app;
 pub use rtic_core::{prelude as mutex_prelude, Exclusive, Mutex};
 pub use rtic_monotonic::{self, Monotonic};
 
+/// module `mutex::prelude` provides `Mutex` and multi-lock variants. Recommended over `mutex_prelude`
+pub mod mutex {
+    pub use rtic_core::prelude;
+}
+
 #[doc(hidden)]
 pub mod export;
 #[doc(hidden)]